Implemented Oteria's search functionality into the homepage display. Updated a lot of styling.
