+++ IT-Wizards-UI/src/App.jsx
@@ -17,7 +17,6 @@ import CartProvider from './components/CartContext';
 import EditProfile from './components/profile/EditProfile';
 import Signup from './components/authentication/Signup';
 import Login from './components/authentication/Login';
-import SearchResultsPage from './pages/SearchResultsPage';
 
 const App = () => {
 
@@ -38,7 +37,6 @@ const App = () => {
         <Route path="edit-profile" element={<EditProfile />} />
         <Route path="/api/users/signup" element={<Signup />} />
         <Route path="/api/users/signin" element={<Login />} />
-        <Route path="/searchResults" element={<SearchResultsPage/>}/>
       </Route>
     )
   );
