+++ IT-Wizards-UI/src/components/AddItemCategoryForm.jsx
@@ -43,7 +43,7 @@ const AddItemCategoryForm = () => {
   };
   return (
     <section className="bg-purple-400">
-      <div className="container m-auto max-w-3xl py-24">
+      <div className="container m-auto max-w-3xl pt-60 pb-96">
         <div className="bg-white px-6 py-8 mb-4 shadow-md rounded-md border m-4 md:m-0">
           <form onSubmit={(e) => onSubmit(e)}>
             <h2 className="text-3xl text-center font-semibold mb-6">
