+++ IT-Wizards-UI/src/components/search/SearchBar.jsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { MagnifyingGlass } from 'phosphor-react';
 import { useNavigate } from 'react-router-dom';
 
-const SearchBar = ({ setResults, setSearchTerm }) => {
+const SearchBar = ({ setResults, setSearchTerm, results }) => {
   const [input, setInput] = useState('');
 
   const fetchSearchResults = (value) => {
@@ -29,19 +29,12 @@ const SearchBar = ({ setResults, setSearchTerm }) => {
     fetchSearchResults(value);
   };
 
-  const handleKeyDown = (e) => {
-    if (e.key === 'Enter') {
-      console.log('test');
-      navigate('/searchResults');
-    }
-  };
-
   return (
     <div className="flex items-center m-auto w-full bg-purple-800 text-green-400">
       <form className="flex items-center max-w-sm mx-auto py-3">
         <label className="sr-only">Search</label>
         <div className="relative w-full">
-          <div className="absolute inset-y-0 start-0 flex items-center ps-3 pointer-events-none"></div>
+          
           <input
             type="text"
             id="search"
@@ -49,13 +42,11 @@ const SearchBar = ({ setResults, setSearchTerm }) => {
             placeholder="Search for an item..."
             value={input}
             onChange={(e) => handleChange(e.target.value)}
-            required
-            onKeyDown={(e) => handleKeyDown(e)}
           />
         </div>
         <button
           type="submit"
-          onClick={()=> navigate('/searchResults')}
+          onClick={()=> navigate(`/items/${results[0].id}`)}
           className="p-2.5 ms-2 text-lg font-medium text-black bg-green-500 rounded-lg border border-green-700 hover:bg-green-600 focus:ring-4 "
         >
           <MagnifyingGlass />
