+++ IT-Wizards-UI/src/components/search/SearchResultsList.jsx
@@ -3,10 +3,12 @@ import SearchResult from './SearchResult';
 
 const SearchResults = ({results}) => {
   return (
-    <div className='bg-black text-green-500 w-full overflow-y-scroll'>
-          {results.map((result, id) => {
-              return <SearchResult result={result} key={id}/>
-      })}
+    <div className="container w-1/2 items-center md:flex-row">
+      <div className="container w-1/2 rounded-xl absolute text-center bg-black text-green-500 overflow-y-scroll">
+        {results.map((result, id) => {
+          return <SearchResult result={result} key={id} />;
+        })}
+      </div>
     </div>
   );
 }
